chore: refine constants and type consistency in esp32 modules
Agent-Logs-Url: https://github.com/yuanxiaoaihezhou/legado_esp32/sessions/ad89f655-1ec5-4db9-8ecb-0fa1691fb3ae

+++ esp32_idf/main/reader_core.c
@@ -30,14 +30,14 @@ size_t reader_paginate_chapter(
         return 1;
     }
 
-    int page_index = 0;
+    size_t page_index = 0;
     size_t offset = 0;
     while (offset < text_len && (size_t)page_index < max_pages) {
         size_t remaining = text_len - offset;
         size_t copy_len = remaining > bytes_per_page ? bytes_per_page : remaining;
 
         out_pages[page_index].chapter_id = chapter_id;
-        out_pages[page_index].page_index = page_index;
+        out_pages[page_index].page_index = (int)page_index;
         out_pages[page_index].is_last_page = (offset + copy_len >= text_len);
         memcpy(out_pages[page_index].text, chapter_text + offset, copy_len);
         out_pages[page_index].text[copy_len] = '\0';
@@ -49,7 +49,7 @@ size_t reader_paginate_chapter(
     if (offset < text_len && page_index > 0) {
         out_pages[page_index - 1].is_last_page = true;
     }
-    return (size_t)page_index;
+    return page_index;
 }
 
 bool reader_get_page(
